backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytest==7.4.3
flake8==6.1.0
alembic==1.12.1
schedule==1.2.1
jinja2==3.1.2
requests==2.31.0
pandas==2.1.4
fpdf2==2.7.6
xlsxwriter==3.1.9
aiohttp==3.9.1
pgvector==0.2.3
pytz==2023.3
